removed redundant std:: typed from habit.
Adjusted iterator::advance() signatures.

Added starting pos for find and rfind (also in doc).

Improved name of an internal method to copy_part, which is what it does.

Moved substr implementation to reflect order of class definition.

Removed code to sum up sizes of parts in bytes() function, would be totally unusable information.

Implemented free functions insert,replace and erase.

Remaining implementation is only toupper/lower/capitalize where the problem is that C++ does not have predicates for characters other than char and wchar_t.

Also remaining are constructors and some other functions which according to ustring.md exist and take a locale&. Problem here (and in general with char) is that C++ has no way portable way to know what the 'system' char encoding that filesystem (and narrow OS calls in general) is. This is only a problem on Windows, and there could be some MS additions for this. Linux always uses UTF-8, AFAIK. Also literals may have another encoding than the system (literal encoding is set when the program is compiled, system encoding when it runs). P1885 allows getting a name and MIB identifier for these encodings but there is nothing to use to convert using those identifiers.